refactor: split as a separate module
